+++ src/main/java/co/simplon/titrepro/isophoto/repository/CategorieRepository.java
@@ -10,7 +10,7 @@ import co.simplon.titrepro.isophoto.model.Categorie;
 public interface CategorieRepository extends JpaRepository<Categorie, String> {
 	
 	@Query("SELECT g FROM Categorie g WHERE g.id LIKE %?1%")
-	  public Categorie findByCategorieId(int id);
+	  public Categorie findByCategorieId(Long id);
 
 	public Optional<Categorie> findByNom(String nom);
